refactor: move markdown editor skeleton loading to proper dynamic import pattern
Replace manual skeleton management in BookProgress with proper Next.js
dynamic import pattern. Split editor into MarkdownEditorInternal.tsx
containing the heavy MDXEditor component, and use dynamic() with loading
option to show skeleton during code-split load. This eliminates content
shift on initial render by ensuring skeleton and editor have identical
dimensions.
